third_party/blink/web_tests/fast/events/no_hover_update_after_layout_change.html

<!DOCTYPE html>
<script src='../../resources/gesture-util.js'></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<style>
#blue {
    background-color: rgb(0, 0, 255);
    position: absolute;
    left: 75px;
    top: 75px;
    height: 100px;
    width: 100px;
}
#blue:hover {
    background-color: rgb(255, 255, 0);
}
</style>

<script>
internals.runtimeFlags.updateHoverFromLayoutChangeAtBeginFrameEnabled = true;

var eventList = [];
var x = 100;
var y = 100;
var addedBlue = false;

function addBlue() {
  document.body.innerHTML += '<div id="blue"></div>';
  var blue = document.getElementById("blue");
  var events = ['mouseover', 'mousemove', 'mouseout', 'mouseenter', 'mouseleave'];
  events.forEach(function (event) {
      blue.addEventListener(event, addMouseEvents);
  });
  addedBlue = true;
}

function checkMouseOverEvent() {
  var result = eventList.join();
  assert_true(result == '');
  assert_equals(getComputedStyle(document.getElementById("blue")).backgroundColor, "rgb(0, 0, 255)");
  assert_true(internals.runtimeFlags.updateHoverFromLayoutChangeAtBeginFrameEnabled)
}

function addMouseEvents(event) {
  eventList.push(event.type);
}

promise_test (async () => {
  document.addEventListener('click', addBlue);
  await mouseClickOn(x, y);
  await waitFor(() => { return addedBlue; });
  await conditionHolds(() => { return addedBlue; });
  checkMouseOverEvent();
}, 'Tests that the mouseover event is not fired and the element has no hover effect when the element underneath the mouse cursor is changed.');
</script>
